refactor: Update imports for AccommodationAmenity, RoomAmenity, and User models to specific class files
